refactor: refine column statistic map serialization
While serializing ColumnStatistics map, excludes columns which is of unsupported type
