+++ src/query/storages/common/index/src/range_index.rs
@@ -247,4 +247,8 @@ pub fn statistics_to_domain(mut stats: Vec<&ColumnStatistics>, data_type: &DataT
     }
 }
 
-impl Index for RangeIndex {}
+impl Index for RangeIndex {
+    fn supported_type(data_type: &DataType) -> bool {
+        databend_storages_common_table_meta::meta::supported_stat_type(data_type)
+    }
+}
