+++ src/query/storages/common/table_meta/Cargo.toml
@@ -20,6 +20,7 @@ databend-common-expression = { workspace = true }
 databend-common-io = { workspace = true }
 databend-common-storage = { path = "../../../../common/storage" }
 enum-as-inner = "0.5"
+log = { workspace = true }
 parquet = { workspace = true }
 rmp-serde = "1.1.1"
 serde = { workspace = true }
