+++ src/query/storages/common/table_meta/src/meta/statistics.rs
@@ -15,6 +15,7 @@
 use std::collections::HashMap;
 
 use databend_common_base::base::uuid::Uuid;
+use databend_common_expression::types::DataType;
 use databend_common_expression::ColumnId;
 
 use crate::meta::ColumnStatistics;
@@ -35,4 +36,16 @@ pub struct BlockSlotDescription {
     // `block_index` mod `num_slots` == `slot_index` indicates that the block should be taken care of by current executor
     // otherwise, the block should be taken care of by other executors
     pub slot: u32,
+}
+
+pub fn supported_stat_type(data_type: &DataType) -> bool {
+    let inner_type = data_type.remove_nullable();
+    matches!(
+        inner_type,
+        DataType::Number(_)
+            | DataType::Date
+            | DataType::Timestamp
+            | DataType::String
+            | DataType::Decimal(_)
+    )
 }
